chore(refactor): move customize site page context
