+++ src/createSchemaCustomization/customizeSitePageContext.ts
@@ -1,0 +1,13 @@
+import { GatsbyNodeCreateSchemaCustomization } from '../../types';
+
+export const customizeSitePageContext: GatsbyNodeCreateSchemaCustomization = async ({ actions }) => {
+  actions.createTypes(`
+    type SitePage implements Node {
+      context: SitePageContext
+    }
+    type SitePageContext {
+      locale: String
+      prefix: String
+    }
+  `);
+};
