+++ types.d.ts
@@ -40,6 +40,7 @@ export type PluginOptions = {
 
 type GatsbyNodeOnCreatePage = NonNullable<GatsbyNode['onCreatePage']>;
 type GatsbyNodeOnCreateNode = NonNullable<GatsbyNode['onCreateNode']>;
+type GatsbyNodeCreateSchemaCustomization = NonNullable<GatsbyNode['createSchemaCustomization']>;
 
 export type OnCreatePage = (args: Parameters<GatsbyNodeOnCreatePage>[0], options?: PluginOptions) => ReturnType<GatsbyNodeOnCreatePage>;
 export type OnCreateNode = (args: Parameters<GatsbyNodeOnCreateNode>[0], options?: PluginOptions) => ReturnType<GatsbyNodeOnCreateNode>;
